PEP 695 functions
